Determiner remakes for PR
